+++ tests/dummy/app/JsonApi/V1/Server.php
@@ -65,6 +65,7 @@ class Server extends BaseServer
         return [
             Comments\CommentSchema::class,
             Images\ImageSchema::class,
+            Phones\PhoneSchema::class,
             Posts\PostSchema::class,
             Tags\TagSchema::class,
             Users\UserSchema::class,