+++ tests/dummy/routes/api.php
@@ -33,7 +33,9 @@ JsonApiRoute::server('v1')
         });
 
         /** Users */
-        $server->resource('users')->only('show')->actions(function ($actions) {
+        $server->resource('users')->only('show')->relationships(function ($relationships) {
+            $relationships->hasOne('phone');
+        })->actions(function ($actions) {
             $actions->get('me');
         });