+++ tests/dummy/tests/Api/V1/Posts/AttachMediaTest.php
@@ -166,12 +166,17 @@ class AttachMediaTest extends TestCase
         $this->assertDatabaseCount('post_video', 1);
     }
 
-    public function testNotAcceptableMediaType(): void
+    /**
+     * @param string $mediaType
+     * @return void
+     * @dataProvider notAcceptableMediaTypeProvider
+     */
+    public function testNotAcceptableMediaType(string $mediaType): void
     {
         $response = $this
             ->actingAs($this->post->author)
             ->jsonApi('posts')
-            ->accept('text/html')
+            ->accept($mediaType)
             ->withData([])
             ->post(url('/api/v1/posts', [$this->post, 'relationships', 'media']));