+++ tests/dummy/tests/Api/V1/Posts/ReadAuthorIdentifierTest.php
@@ -137,10 +137,16 @@ class ReadAuthorIdentifierTest extends TestCase
         );
     }
 
-    public function testInvalidMediaType(): void
+    /**
+     * @param string $mediaType
+     * @return void
+     * @dataProvider notAcceptableMediaTypeProvider
+     */
+    public function testNotAcceptableMediaType(string $mediaType): void
     {
-        $this->jsonApi()
-            ->accept('text/html')
+        $this
+            ->jsonApi()
+            ->accept($mediaType)
             ->get(url('/api/v1/posts', [$this->post, 'relationships', 'author']))
             ->assertStatus(406);
     }