+++ tests/dummy/tests/Api/V1/Posts/ReadMediaTest.php
@@ -168,10 +168,15 @@ class ReadMediaTest extends TestCase
         $response->assertFetchedManyExact($expected);
     }
 
-    public function testInvalidMediaType(): void
+    /**
+     * @param string $mediaType
+     * @return void
+     * @dataProvider notAcceptableMediaTypeProvider
+     */
+    public function testNotAcceptableMediaType(string $mediaType): void
     {
         $this->jsonApi()
-            ->accept('text/html')
+            ->accept($mediaType)
             ->get(url('/api/v1/posts', [$this->post, 'media']))
             ->assertStatus(406);
     }