+++ tests/dummy/tests/Api/V1/Posts/UpdateTagsTest.php
@@ -149,12 +149,17 @@ class UpdateTagsTest extends TestCase
         $this->assertDatabaseCount('taggables', 2);
     }
 
-    public function testNotAcceptableMediaType(): void
+    /**
+     * @param string $mediaType
+     * @return void
+     * @dataProvider notAcceptableMediaTypeProvider
+     */
+    public function testNotAcceptableMediaType(string $mediaType): void
     {
         $response = $this
             ->actingAs($this->post->author)
             ->jsonApi('posts')
-            ->accept('text/html')
+            ->accept($mediaType)
             ->withData([])
             ->patch(url('/api/v1/posts', [$this->post, 'relationships', 'tags']));